+++ src/editor/parser/StateRender.js
@@ -1,8 +1,11 @@
 import { LOWERCASE_TAGS, CLASS_OR_ID } from '../config'
-import { conflict, isLengthEven, isEven, getIdWithoutSet, loadImage, getImageSrc } from '../utils'
+import { conflict, isLengthEven, union, isEven, getIdWithoutSet, loadImage, getImageSrc } from '../utils'
 import { insertAfter, operateClassName } from '../utils/domManipulate.js'
 import { tokenizer } from './parse'
 import { validEmoji } from '../emojis'
+
+// for test
+window.tokenizer = tokenizer
 
 const snabbdom = require('snabbdom')
 const patch = snabbdom.init([ // Init patch function with chosen modules
@@ -45,11 +48,15 @@ class StateRender {
     return outerClass || (this.checkConflicted(block, token, cursor) ? CLASS_OR_ID['AG_GRAY'] : CLASS_OR_ID['AG_HIDE'])
   }
 
+  getHighlightClassName (active) {
+    return active ? CLASS_OR_ID['AG_HIGHLIGHT'] : CLASS_OR_ID['AG_SELECTION']
+  }
+
   /**
    * [render]: 2 steps:
    * render vdom
    */
-  render (blocks, cursor, activeBlocks) {
+  render (blocks, cursor, activeBlocks, matches) {
     const selector = `${LOWERCASE_TAGS.div}#${CLASS_OR_ID['AG_EDITOR_ID']}`
 
     const renderBlock = block => {
@@ -115,8 +122,10 @@ class StateRender {
 
         return h(blockSelector, data, block.children.map(child => renderBlock(child)))
       } else {
+        // highlight search key in block
+        const highlights = matches.filter(m => m.key === block.key)
         let children = block.text
-          ? tokenizer(block.text).reduce((acc, token) => {
+          ? tokenizer(block.text, highlights).reduce((acc, token) => {
             const chunk = this[token.type](h, cursor, block, token)
             return Array.isArray(chunk) ? [...acc, ...chunk] : [...acc, chunk]
           }, [])
@@ -184,110 +193,211 @@ class StateRender {
   }
 
   hr (h, cursor, block, token, outerClass) {
+    const { start, end } = token.range
+    const content = this.highlight(h, block, start, end, token)
     return [
-      h(`span.${CLASS_OR_ID['AG_GRAY']}.${CLASS_OR_ID['AG_REMOVE']}`, token.marker)
+      h(`span.${CLASS_OR_ID['AG_GRAY']}.${CLASS_OR_ID['AG_REMOVE']}`, content)
     ]
   }
 
   header (h, cursor, block, token, outerClass) {
     const className = this.getClassName(outerClass, block, token, cursor)
+    const { start, end } = token.range
+    const content = this.highlight(h, block, start, end, token)
     return [
-      h(`span.${className}.${CLASS_OR_ID['AG_REMOVE']}`, token.marker)
+      h(`span.${className}.${CLASS_OR_ID['AG_REMOVE']}`, content)
     ]
   }
 
   ['code_fense'] (h, cursor, block, token, outerClass) {
+    const { start, end } = token.range
+    const { marker } = token
+
+    const markerContent = this.highlight(h, block, start, start + marker.length, token)
+    const content = this.highlight(h, block, start + marker.length, end, token)
+
     return [
-      h(`span.${CLASS_OR_ID['AG_GRAY']}`, token.marker),
-      h(`span.${CLASS_OR_ID['AG_LANGUAGE']}`, token.content)
+      h(`span.${CLASS_OR_ID['AG_GRAY']}`, markerContent),
+      h(`span.${CLASS_OR_ID['AG_LANGUAGE']}`, content)
     ]
   }
 
   backlash (h, cursor, block, token, outerClass) {
     const className = this.getClassName(outerClass, block, token, cursor)
+    const { start, end } = token.range
+    const content = this.highlight(h, block, start, end, token)
+
     return [
-      h(`span.${className}.${CLASS_OR_ID['AG_REMOVE']}`, token.marker)
+      h(`span.${className}.${CLASS_OR_ID['AG_REMOVE']}`, content)
     ]
   }
 
   ['inline_code'] (h, cursor, block, token, outerClass) {
     const className = this.getClassName(outerClass, block, token, cursor)
+    const { marker } = token
+    const { start, end } = token.range
+
+    const startMarker = this.highlight(h, block, start, start + marker.length, token)
+    const endMarker = this.highlight(h, block, end - marker.length, end, token)
+    const content = this.highlight(h, block, start + marker.length, end - marker.length, token)
+
     return [
-      h(`span.${className}.${CLASS_OR_ID['AG_REMOVE']}`, token.marker),
-      h('code', token.content),
-      h(`span.${className}.${CLASS_OR_ID['AG_REMOVE']}`, token.marker)
+      h(`span.${className}.${CLASS_OR_ID['AG_REMOVE']}`, startMarker),
+      h('code', content),
+      h(`span.${className}.${CLASS_OR_ID['AG_REMOVE']}`, endMarker)
     ]
   }
+  // change text to highlight vdom
+  highlight (h, block, rStart, rEnd, token) {
+    const { text } = block
+    const { highlights } = token
+    let result = []
+    let unions = []
+    let pos = rStart
 
-  text (h, cursor, block, token) {
-    return token.content
+    if (highlights) {
+      for (const light of highlights) {
+        const un = union({ start: rStart, end: rEnd }, light)
+        if (un) unions.push(un)
+      }
+    }
+
+    if (unions.length) {
+      for (const u of unions) {
+        const { start, end, active } = u
+        const className = this.getHighlightClassName(active)
+
+        if (pos < start) {
+          result.push(text.substring(pos, start))
+        }
+
+        result.push(h(`span.${className}`, text.substring(start, end)))
+        pos = end
+      }
+      if (pos < rEnd) {
+        result.push(block.text.substring(pos, rEnd))
+      }
+    } else {
+      result = [ text.substring(rStart, rEnd) ]
+    }
+
+    return result
   }
-
+  // render token of text type to vdom.
+  text (h, cursor, block, token) {
+    const { start, end } = token.range
+    return this.highlight(h, block, start, end, token)
+  }
+  // render token of emoji to vdom
   emoji (h, cursor, block, token, outerClass) {
+    const { start: rStart, end: rEnd } = token.range
     const className = this.getClassName(outerClass, block, token, cursor)
     const validation = validEmoji(token.content)
     const finalClass = validation ? className : CLASS_OR_ID['AG_WARN']
+    const CONTENT_CLASSNAME = `span.${finalClass}.${CLASS_OR_ID['AG_EMOJI_MARKED_TEXT']}`
+    let startMarkerCN = `span.${finalClass}.${CLASS_OR_ID['AG_EMOJI_MARKER']}`
+    let endMarkerCN = startMarkerCN
+    let content = token.content
+    let pos = rStart + token.marker.length
+
+    if (token.highlights && token.highlights.length) {
+      content = []
+      for (const light of token.highlights) {
+        let { start, end, active } = light
+        const HIGHLIGHT_CLASSNAME = this.getHighlightClassName(active)
+        if (start === rStart) {
+          startMarkerCN += `.${HIGHLIGHT_CLASSNAME}`
+          start++
+        }
+        if (end === rEnd) {
+          endMarkerCN += `.${HIGHLIGHT_CLASSNAME}`
+          end--
+        }
+        if (pos < start) {
+          content.push(block.text.substring(pos, start))
+        }
+        if (start < end) {
+          content.push(h(`span.${HIGHLIGHT_CLASSNAME}`, block.text.substring(start, end)))
+        }
+        pos = end
+      }
+      if (pos < rEnd - token.marker.length) {
+        content.push(block.text.substring(pos, rEnd - 1))
+      }
+    }
+
     const emojiVdom = validation
-      ? h(`span.${finalClass}.${CLASS_OR_ID['AG_EMOJI_MARKED_TEXT']}`, { dataset: { emoji: validation.emoji } }, token.content)
-      : h(`span.${finalClass}.${CLASS_OR_ID['AG_EMOJI_MARKED_TEXT']}`, token.content)
+      ? h(CONTENT_CLASSNAME, {
+        dataset: {
+          emoji: validation.emoji
+        }
+      }, content)
+      : h(CONTENT_CLASSNAME, content)
 
     return [
-      h(`span.${finalClass}.${CLASS_OR_ID['AG_EMOJI_MARKER']}`, token.marker),
+      h(startMarkerCN, token.marker),
       emojiVdom,
-      h(`span.${finalClass}.${CLASS_OR_ID['AG_EMOJI_MARKER']}`, token.marker)
+      h(endMarkerCN, token.marker)
     ]
   }
 
   // render factory of `del`,`em`,`strong`
   delEmStrongFac (type, h, cursor, block, token, outerClass) {
     const className = this.getClassName(outerClass, block, token, cursor)
+    const COMMON_MARKER = `span.${className}.${CLASS_OR_ID['AG_REMOVE']}`
+    const { marker } = token
+    const { start, end } = token.range
+    const backlashStart = end - marker.length - token.backlash.length
+    const content = [
+      ...token.children.reduce((acc, to) => {
+        const chunk = this[to.type](h, cursor, block, to, className)
+        return Array.isArray(chunk) ? [...acc, ...chunk] : [...acc, chunk]
+      }, []),
+      ...this.backlashInToken(token.backlash, className, backlashStart, token)
+    ]
+    const startMarker = this.highlight(h, block, start, start + marker.length, token)
+    const endMarker = this.highlight(h, block, end - marker.length, end, token)
 
     if (isLengthEven(token.backlash)) {
       return [
-        h(`span.${className}.${CLASS_OR_ID['AG_REMOVE']}`, token.marker),
-        h(type, [
-          ...token.children.reduce((acc, to) => {
-            const chunk = this[to.type](h, cursor, block, to, className)
-            return Array.isArray(chunk) ? [...acc, ...chunk] : [...acc, chunk]
-          }, []),
-          ...this.backlashInToken(token.backlash, className)
-        ]),
-        h(`span.${className}.${CLASS_OR_ID['AG_REMOVE']}`, token.marker)
+        h(COMMON_MARKER, startMarker),
+        h(type, content),
+        h(COMMON_MARKER, endMarker)
       ]
     } else {
       return [
-        token.marker,
-        ...token.children.reduce((acc, to) => {
-          const chunk = this[to.type](h, cursor, block, to, className)
-          return Array.isArray(chunk) ? [...acc, ...chunk] : [...acc, chunk]
-        }, []),
-        ...this.backlashInToken(token.backlash, className),
-        token.marker
+        ...startMarker,
+        ...content,
+        ...endMarker
       ]
     }
   }
-
-  backlashInToken (backlashes, outerClass) {
+  // TODO HIGHLIGHT
+  backlashInToken (backlashes, outerClass, start, token) {
+    const { highlights = [] } = token
     const chunks = backlashes.split('')
     const len = chunks.length
     const result = []
     let i
 
     for (i = 0; i < len; i++) {
+      const chunk = chunks[i]
+      const light = highlights.filter(light => union({ start: start + i, end: start + i + 1 }, light))
+      let selector = 'span'
+      if (light.length) {
+        const className = this.getHighlightClassName(light[0].active)
+        selector += `.${className}`
+      }
       if (isEven(i)) {
         result.push(
-          h(`span.${outerClass}`, chunks[i])
+          h(`${selector}.${outerClass}`, chunk)
         )
       } else {
         result.push(
-          h(`span.${CLASS_OR_ID['AG_BACKLASH']}`, chunks[i])
+          h(`${selector}.${CLASS_OR_ID['AG_BACKLASH']}`, chunk)
         )
       }
     }
-
-    // result.push(
-    //   h(`span.${CLASS_OR_ID['AG_BUG']}`) // the extral a tag for fix bug
-    // )
 
     return result
   }
@@ -295,6 +405,23 @@ class StateRender {
   image (h, cursor, block, token, outerClass) {
     const className = this.getClassName(outerClass, block, token, cursor)
     const imageClass = CLASS_OR_ID['AG_IMAGE_MARKED_TEXT']
+
+    const {
+      start,
+      end
+    } = token.range
+    const titleContent = this.highlight(h, block, start, start + 2 + token.title.length, token)
+    const srcContent = this.highlight(
+      h, block,
+      start + 2 + token.title.length + token.backlash.first.length,
+      start + 2 + token.title.length + token.backlash.first.length + 2 + token.src.length,
+      token
+    )
+    const firstBracketContent = this.highlight(h, block, start, start + 2, token)
+    const lastBracketContent = this.highlight(h, block, end - 1, end, token)
+
+    const firstBacklashStart = start + 2 + token.title.length
+    const secondBacklashStart = end - 1 - token.backlash.second.length
 
     if (isLengthEven(token.backlash.first) && isLengthEven(token.backlash.second)) {
       let id
@@ -347,11 +474,11 @@ class StateRender {
       }
 
       const children = [
-        `![${token.title}`,
-        ...this.backlashInToken(token.backlash.first, className),
-        `](${token.src}`,
-        ...this.backlashInToken(token.backlash.second, className),
-        ')'
+        ...titleContent,
+        ...this.backlashInToken(token.backlash.first, className, firstBacklashStart, token),
+        ...srcContent,
+        ...this.backlashInToken(token.backlash.second, className, secondBacklashStart, token),
+        ...lastBracketContent
       ]
 
       return isSuccess
@@ -362,27 +489,29 @@ class StateRender {
         : [h(selector, children)]
     } else {
       return [
-        '![',
+        ...firstBracketContent,
         ...token.children.reduce((acc, to) => {
           const chunk = this[to.type](h, cursor, block, to, className)
           return Array.isArray(chunk) ? [...acc, ...chunk] : [...acc, chunk]
         }, []),
-        ...this.backlashInToken(token.backlash.first, className),
-        '](',
-        token.src,
-        ...this.backlashInToken(token.backlash.second, className),
-        ')'
+        ...this.backlashInToken(token.backlash.first, className, firstBacklashStart, token),
+        ...srcContent,
+        ...this.backlashInToken(token.backlash.second, className, secondBacklashStart, token),
+        ...lastBracketContent
       ]
     }
   }
-
+  // render auto_link to vdom
   ['auto_link'] (h, cursor, block, token, outerClass) {
+    const { start, end } = token.range
+    const content = this.highlight(h, block, start, end, token)
+
     return [
       h('a', {
         props: {
           href: token.href
         }
-      }, token.href)
+      }, content)
     ]
   }
 
@@ -390,24 +519,50 @@ class StateRender {
   link (h, cursor, block, token, outerClass) {
     const className = this.getClassName(outerClass, block, token, cursor)
     const linkClassName = className === CLASS_OR_ID['AG_HIDE'] ? className : CLASS_OR_ID['AG_LINK_IN_BRACKET']
+    const { start, end } = token.range
+    const firstMiddleBracket = this.highlight(h, block, start, start + 3, token)
+
+    const firstBracket = this.highlight(h, block, start, start + 1, token)
+    const middleBracket = this.highlight(
+      h, block,
+      start + 1 + token.anchor.length + token.backlash.first.length,
+      start + 1 + token.anchor.length + token.backlash.first.length + 2,
+      token
+    )
+    const hrefContent = this.highlight(
+      h, block,
+      start + 1 + token.anchor.length + token.backlash.first.length + 2,
+      start + 1 + token.anchor.length + token.backlash.first.length + 2 + token.href.length,
+      token
+    )
+    const middleHref = this.highlight(
+      h, block, start + 1 + token.anchor.length + token.backlash.first.length,
+      block, start + 1 + token.anchor.length + token.backlash.first.length + 2 + token.href.length,
+      token
+    )
+
+    const lastBracket = this.highlight(h, block, end - 1, end, token)
+
+    const firstBacklashStart = start + 1 + token.anchor.length
+    const secondBacklashStart = end - 1 - token.backlash.second.length
 
     if (isLengthEven(token.backlash.first) && isLengthEven(token.backlash.second)) {
       if (!token.children.length && !token.backlash.first) { // no-text-link
         return [
-          h(`span.${CLASS_OR_ID['AG_GRAY']}.${CLASS_OR_ID['AG_REMOVE']}`, '[]('),
+          h(`span.${CLASS_OR_ID['AG_GRAY']}.${CLASS_OR_ID['AG_REMOVE']}`, firstMiddleBracket),
           h(`a.${CLASS_OR_ID['AG_NOTEXT_LINK']}`, {
             props: {
               href: token.href + encodeURI(token.backlash.second)
             }
           }, [
-            token.href,
-            ...this.backlashInToken(token.backlash.second, className)
+            ...hrefContent,
+            ...this.backlashInToken(token.backlash.second, className, secondBacklashStart, token)
           ]),
-          h(`span.${CLASS_OR_ID['AG_GRAY']}.${CLASS_OR_ID['AG_REMOVE']}`, ')')
+          h(`span.${CLASS_OR_ID['AG_GRAY']}.${CLASS_OR_ID['AG_REMOVE']}`, lastBracket)
         ]
       } else { // has children
         return [
-          h(`span.${className}.${CLASS_OR_ID['AG_REMOVE']}`, '['),
+          h(`span.${className}.${CLASS_OR_ID['AG_REMOVE']}`, firstBracket),
           h('a', {
             dataset: {
               href: token.href + encodeURI(token.backlash.second)
@@ -417,27 +572,27 @@ class StateRender {
               const chunk = this[to.type](h, cursor, block, to, className)
               return Array.isArray(chunk) ? [...acc, ...chunk] : [...acc, chunk]
             }, []),
-            ...this.backlashInToken(token.backlash.first, className)
+            ...this.backlashInToken(token.backlash.first, className, firstBacklashStart, token)
           ]),
-          h(`span.${className}.${CLASS_OR_ID['AG_REMOVE']}`, ']('),
+          h(`span.${className}.${CLASS_OR_ID['AG_REMOVE']}`, middleBracket),
           h(`span.${linkClassName}.${CLASS_OR_ID['AG_REMOVE']}`, [
-            token.href,
-            ...this.backlashInToken(token.backlash.second, className)
+            ...hrefContent,
+            ...this.backlashInToken(token.backlash.second, className, secondBacklashStart, token)
           ]),
-          h(`span.${className}.${CLASS_OR_ID['AG_REMOVE']}`, ')')
+          h(`span.${className}.${CLASS_OR_ID['AG_REMOVE']}`, lastBracket)
         ]
       }
     } else {
       return [
-        '[',
+        ...firstBracket,
         ...token.children.reduce((acc, to) => {
           const chunk = this[to.type](h, cursor, block, to, className)
           return Array.isArray(chunk) ? [...acc, ...chunk] : [...acc, chunk]
         }, []),
-        ...this.backlashInToken(token.backlash.first, className),
-        `](${token.href}`,
-        ...this.backlashInToken(token.backlash.second, className),
-        ')'
+        ...this.backlashInToken(token.backlash.first, className, firstBacklashStart, token),
+        ...middleHref,
+        ...this.backlashInToken(token.backlash.second, className, secondBacklashStart, token),
+        ...lastBracket
       ]
     }
   }